requirements.txt
Flask        
Flask-Assets
Flask-Babel
Flask-Script
httplib2==0.9.2       
Jinja2==2.9.4            
oauth2==1.9.0.post1   
python-editor==1.0.3  
requests==2.12.4
webassets==0.12.0
Werkzeug==0.11.15     
gunicorn==19.6.0
line-bot-sdk
mysqlclient
